backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "FastAPI backend for FaithChat"
requires-python = ">=3.10"
dependencies = [
  "fastapi",
  "uvicorn[standard]",
  "python-multipart",
  "pydantic>=2",
  "python-jose[cryptography]",
  "passlib[bcrypt]",
  "openai>=1.0.0",
  "httpx",
  "starlette",
  "pyjwt"
]

[tool.uvicorn]
host = "0.0.0.0"
port = 8000
